tech(api-1d): move assessment get next challenge route
