+++ api/lib/application/assessments/index.js
@@ -76,22 +76,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/pix1d/assessments/{id}/next',
-      config: {
-        pre: [{ method: securityPreHandlers.checkPix1dActivated }],
-        auth: false,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.assessmentId,
-          }),
-        },
-        handler: assessmentController.getNextChallengeForPix1d,
-        notes: ["- Récupération de la question suivante pour l'évaluation de mission donnée"],
-        tags: ['api'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/assessments/{id}',
       config: {
         auth: false,
