+++ api/src/school/routes.js
@@ -1,0 +1,5 @@
+import * as assessment from './application/assessments/assessment-route.js';
+
+const schoolRoutes = [assessment];
+
+export { schoolRoutes };
