+++ api/src/school/shared/usecases/index.js
@@ -1,0 +1,30 @@
+import { fileURLToPath } from 'node:url';
+import { dirname, join } from 'node:path';
+
+import * as activityAnswerRepository from '../../../../lib/infrastructure/repositories/activity-answer-repository.js';
+import * as activityRepository from '../../../../lib/infrastructure/repositories/activity-repository.js';
+import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
+import * as challengeRepository from '../../../../lib/infrastructure/repositories/challenge-repository.js';
+
+import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
+
+const dependencies = {
+  activityAnswerRepository,
+  activityRepository,
+  assessmentRepository,
+  challengeRepository,
+};
+
+const path = dirname(fileURLToPath(import.meta.url));
+
+const usecasesWithoutInjectedDependencies = {
+  ...(await importNamedExportsFromDirectory({
+    path: join(path, '../../domain/usecases'),
+    ignoredFileNames: ['index.js'],
+  })),
+};
+
+const usecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
+
+export { usecases };
