+++ api/tests/school/unit/application/assessments/assessment-controller_test.js
@@ -1,8 +1,6 @@
 import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { assessmentController } from '../../../../../lib/application/assessments/assessment-controller.js';
-import { usecases } from '../../../../../lib/domain/usecases/index.js';
-import * as events from '../../../../../lib/domain/events/index.js';
-
+import { assessmentController } from '../../../../../src/school/application/assessments/assessment-controller.js';
+import { usecases } from '../../../../../src/school/shared/usecases/index.js';
 describe('Unit | Controller | assessment-controller', function () {
   describe('#getNextChallengeForPix1d', function () {
     it('should call the expected usecase', async function () {
