+++ api/tests/unit/application/assessments/assessment-controller_test.js
@@ -105,32 +105,6 @@ describe('Unit | Controller | assessment-controller', function () {
     });
   });
 
-  describe('#getNextChallengeForPix1d', function () {
-    it('should call the expected usecase', async function () {
-      const assessmentId = 104974;
-      const challenge = { id: 'rec1', instruction: '1st challenge for Pix1d' };
-      const challengeSerializerStub = { serialize: sinon.stub() };
-      challengeSerializerStub.serialize.resolves(challenge);
-
-      // given
-      const request = {
-        params: {
-          id: assessmentId,
-        },
-      };
-
-      sinon.stub(usecases, 'getNextChallengeForPix1d').withArgs({ assessmentId }).resolves(challenge);
-
-      // when
-      const result = await assessmentController.getNextChallengeForPix1d(request, hFake, {
-        challengeSerializer: challengeSerializerStub,
-      });
-
-      // then
-      expect(result).to.be.equal(challenge);
-    });
-  });
-
   describe('#completeAssessment', function () {
     let domainTransaction, assessmentId, assessment, assessmentCompletedEvent;
