+++ api/tests/unit/application/assessments/index_test.js
@@ -4,7 +4,6 @@ import { assessmentAuthorization } from '../../../../lib/application/preHandlers
 import { assessmentController } from '../../../../lib/application/assessments/assessment-controller.js';
 import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import * as moduleUnderTest from '../../../../lib/application/assessments/index.js';
-import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
 
 describe('Unit | Application | Router | assessment-router', function () {
   describe('POST /api/assessments', function () {
@@ -66,34 +65,6 @@ describe('Unit | Application | Router | assessment-router', function () {
 
       // then
       expect(response.statusCode).to.equal(200);
-    });
-  });
-
-  describe('GET /api/pix1d/assessments/{id}/next', function () {
-    it('should return 200', async function () {
-      // given
-      sinon.stub(assessmentController, 'getNextChallengeForPix1d').returns('ok');
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/pix1d/assessments/1/next');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 400', async function () {
-      // given
-      sinon.stub(assessmentController, 'getNextChallengeForPix1d').rejects(new AssessmentEndedError());
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/pix1d/assessments/1/next');
-
-      // then
-      expect(response.statusCode).to.equal(400);
     });
   });
